Refactor vista
