+++ src/main/java/org/jugnicaragua/app/pico/vista/catalogo/CuentaForm.java
@@ -1,5 +1,6 @@
-package org.jugnicaragua.app.pico.vista;
+package org.jugnicaragua.app.pico.vista.catalogo;
 
+import org.jugnicaragua.app.pico.vista.menu.MenuPrincipalAcciones;
 import org.jugnicaragua.app.pico.dao.CuentaTipoImpl;
 import org.jugnicaragua.app.pico.modelo.CuentaTipo;
 import org.jugnicaragua.app.pico.vista.utils.ButtonContructor;
